Altering migrations to be less finicky

+++ src/services/database/migrations/20190317203731_inventory-json.js
@@ -1,9 +1,14 @@
 // Refactor shitty, insecure CSV inventory system with a legitimate data structure
 exports.up = knex => {
-  return knex.schema.table('users', table => {
-    table.dropColumn('inventory');
-    table.json('inventory').defaultTo('[]');
-  });
+  return knex.schema
+    .hasColumn('users', 'inventory')
+    .then(inventoryColumnExists => {
+      return knex.schema.table('users', table => {
+        if (inventoryColumnExists) {
+          table.dropColumn('inventory');
+        }
+      });
+    });
 };
 
 exports.down = knex => {
